[AdminPage] Add store for authentication and add history for some files
