+++ webapp/admin/src/api/http.ts
@@ -6,14 +6,15 @@
  * History
  * ============================================================================
  * Number | Date(YYYYMMDD) | Description
- * -------|----------------|----------------------------------------------------------------
+ * -------|----------------|----------------------------------------------------
  *      1 |   2025-09-02   | Initial version
+ *      2 |   2026-03-20   | Add store for authentication
  * 
  * ============================================================================
  */
 
 import axios from 'axios';
-import { authStore } from './authen/auth';
+import { useAuthStore } from '../stores/authen/auth'
 
 const http = axios.create({
   baseURL: import.meta.env.VITE_API_URL,
@@ -22,6 +23,7 @@ const http = axios.create({
 });
 
 http.interceptors.request.use((config) => {
+  const authStore = useAuthStore();
   const store = authStore.getToken();
   if (store != null && store.token) { // check token is existed?
     config.headers = config.headers || {};
